test: split deferred delivery race coverage
Keep the deferred handler regression suite under the file-size gate by isolating real-store race tests. Make the default Claude foreground assertion independent of environment-specific launch flags while retaining exact custom argv coverage.
